fix(gateway): drop httplib from the plugin-facing DTO/provider headers
The typed DTO contract made dto/operations.hpp pull http/alternate_status.hpp ->
http/typed_router.hpp -> <httplib.h>, leaking httplib into every plugin that includes a
provider interface (e.g. operation_provider.hpp). Plugins built against the installed
gateway (build-farm / Docker topology, where the vendored httplib is not on the include
path) then failed with 'httplib.h: No such file'.

Extract the httplib-free handler-result vocabulary (Result, NoContent, Forwarded,
ValidatorResult, ResponseAttachments) into a new leaf header http/handler_result.hpp;
typed_router.hpp re-exports it and keeps only TypedRequest (the sole httplib-dependent
type); alternate_status.hpp includes the leaf. Plugins are now httplib-free across the
.so boundary. No ABI, wire, or behaviour change.
